requirement.txt
flask 
gunicorn
python-binance